% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_stats_corpus.list.R
\name{add.lines.metrics.to.corpus.files}
\alias{add.lines.metrics.to.corpus.files}
\title{Add comments and codeslines metrics}
\usage{
add.lines.metrics.to.corpus.files(corpus, .verbose = T)
}
\arguments{
\item{corpus}{\code{corpus.list} - \code{list} of dataframes created by \code{construct_corpus()}.}

\item{.verbose}{\code{logical} - Add messages if \code{TRUE}.}
}
\value{
A dataframe of class 'corpus.list'.
}
\description{
Add comments and codeslines metrics
}
