% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_functions_list_to_corpus.R
\name{add_functions_list_to_corpus}
\alias{add_functions_list_to_corpus}
\title{Compute a function list from the codes df of a corpus.list object}
\usage{
add_functions_list_to_corpus(
  corpus,
  lang_dictionnary,
  .verbose = F,
  sep = " ",
  fn_to_exclude = NULL
)
}
\arguments{
\item{corpus}{A \code{corpus.list} object with at least the 'codes' dataframe}

\item{lang_dictionnary}{\code{list} A language patterns dictionnary specific to one language}

\item{.verbose}{\code{logical}, default = \code{FALSE}. If set to \code{TRUE}, show a progress bar when extracting content}

\item{sep}{\code{character}, default = \code{' '} A character chain of separator, in order to concatenate lines}

\item{fn_to_exclude}{\code{character} A vector of values that will not be returned such as a match.}
}
\value{
a \code{list} of class \code{corpus.list} with exposed functions names and file path
}
\description{
Compute a function list from the codes df of a corpus.list object
}
