% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_network.R
\name{compute_similarity_network}
\alias{compute_similarity_network}
\title{Compute a similarity network}
\usage{
compute_similarity_network(df, group_col, text_col)
}
\arguments{
\item{df}{\code{data.frame} - A data.frame with a vector of char and a group col}

\item{group_col}{\code{character} - Name of the group vector}

\item{text_col}{\code{character} - Name of the text column vector, i.e. the vector used for compute similarity}
}
\description{
Based on a lines to lines Levenchtein distance (\% of similarity) between lines.
Assuming lines are from several files. Used for adding stats to corpus (methods_stats)
}
