% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{.construct.corpus.lines}
\alias{.construct.corpus.lines}
\title{Construct a corpus.lines dataframe}
\usage{
.construct.corpus.lines(df)
}
\arguments{
\item{df}{A dataframe containing a 'line_text' column.}
}
\value{
A dataframe of class 'corpus.lines'.
}
\description{
Construct a corpus.lines dataframe
}
