% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{.construct.corpus.list}
\alias{.construct.corpus.list}
\title{Create a corpus.list from a list or append a dataframe to a corpus.list}
\usage{
.construct.corpus.list(
  corpus,
  df_to_add = NULL,
  names_of_df_to_add = NULL,
  ...
)
}
\arguments{
\item{corpus}{\code{corpus.list} \code{list} of dataframes created by \code{construct_corpus()}.}

\item{df_to_add}{\code{list} of \code{data.frame} or a single \code{data.frame} to append to the corpus.list. If data.frame(s) without a name are provided, names should be provided with \code{names_of_df_to_add}}

\item{names_of_df_to_add}{\code{character} vector of names to append to the corpus.list, if anonymous materials is provided (unamed df or list of unamed df).}

\item{...}{Additionnal attributes (e.g., keep trace of the languages analyzed).}
}
\value{
A dataframe of class 'corpus.list'.
}
\description{
Create a corpus.list from a list or append a dataframe to a corpus.list
}
