% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{.construct.nodelist}
\alias{.construct.nodelist}
\title{Construct a corpus.nodelist dataframe}
\usage{
.construct.nodelist(df)
}
\arguments{
\item{df}{A dataframe containing a 'line_text' column.}
}
\value{
A dataframe of class 'corpus.nodelist'.
}
\description{
Construct a corpus.nodelist dataframe
}
