% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_char_at_positions.R
\name{escape_char_at_positions}
\alias{escape_char_at_positions}
\title{Escape characters at specific positions}
\usage{
escape_char_at_positions(txt, positions, escaping_prefix = "\\\\\\\\")
}
\arguments{
\item{txt}{The text to process (character).}

\item{positions}{The positions at which to escape characters.}

\item{escaping_prefix}{The prefix to escape the characters with.}
}
\value{
The processed text with escaped characters at specific positions.
}
\description{
Escape characters at specific positions
}
