% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_unescaped_chars.R
\name{escape_unescaped_chars}
\alias{escape_unescaped_chars}
\title{Escape char if unescaped, default to incorrectly enclosed characters}
\usage{
escape_unescaped_chars(
  txt,
  special_char_to_escape = c("(", ")", "[", "]", "{", "}"),
  no_escaping_if_begin_of_chain = FALSE,
  escaping_prefix = "\\\\\\\\"
)
}
\arguments{
\item{txt}{The text to process.}

\item{special_char_to_escape}{Characters to escape.}

\item{no_escaping_if_begin_of_chain}{\code{logical}, default = \code{FALSE} Logical flag. Avoid escape begining of a character chain}

\item{escaping_prefix}{\code{character}, default = \code{"\\\\\\\\"} The prefix to escape the characters with (also used for avoid to escape).}
}
\value{
The processed text with the correctly escaped characters.
}
\description{
Escape char if unescaped, default to incorrectly enclosed characters
}
