% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fn_content.R
\name{extract_fn_content}
\alias{extract_fn_content}
\title{Extract parameters and code for each functions of the functions df of a corpus.list}
\usage{
extract_fn_content(corpus, lang_dictionnary)
}
\arguments{
\item{corpus}{A \code{corpus.list} object with a \code{functions} \code{data.frame}}

\item{lang_dictionnary}{A standard language patterns dictionnary for a language (i.e. params fn_regex_params_after_names is needed)}
}
\description{
Extract parameters and code for each functions of the functions df of a corpus.list
}
