% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_igraph_egonetwork.R
\name{filter_igraph_egonetwork}
\alias{filter_igraph_egonetwork}
\title{Extract an ego-subgraph based on a regex pattern}
\usage{
filter_igraph_egonetwork(
  graph_igraph,
  regex,
  degrees_in_filter = 2,
  degrees_out_filter = 2
)
}
\arguments{
\item{graph_igraph}{An igraph object representing the graph.}

\item{regex}{\code{character} A character string representing the regex pattern to filter nodes.}

\item{degrees_in_filter}{\code{integer}, default = \code{2} An integer specifying the depth of inward reachable nodes (default is 2).}

\item{degrees_out_filter}{\code{integer}, default = \code{2} An integer specifying the depth of outward reachable nodes (default is 2).}
}
\value{
An igraph object representing the filtered subgraph.
}
\description{
Given a \code{igraph} object and a pattern (regex), this function will return
an ego-subgraph of these node(s), within specified depths from them.
If the graph is directed, the nodes are filtered as an ego-network \emph{within specified depths}
}
\examples{
g <- igraph::make_ring(10, directed = TRUE)
igraph::V(g)$name <- as.character(1:10)
subg_2_lvl_out <- filter_igraph_egonetwork(g,  regex= "1$", 0, 2)
plot(subg_2_lvl_out)
}
