% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_escaping.R
\name{fix_escaping}
\alias{fix_escaping}
\title{fix_escaping ensure correct escaping of special characters in a string}
\usage{
fix_escaping(text, special_chars, num_escapes = 2)
}
\arguments{
\item{text}{A character string to process.}

\item{special_chars}{A character vector of special characters that need escaping.}

\item{num_escapes}{Integer. Required number of \verb{\\} before special character.}
}
\value{
A corrected string with the right number of escape characters.
}
\description{
This function adds the designated number of escape characters (\verb{\\}) before special characters.
}
\examples{
codexplor::fix_escaping("This \\\\\\\\(is\\\\) (a\\\\\\\\\\\\) test?"
, special_chars = c("(", ")", "?"), num_escapes = 4)
# Returns: "This \\\\(is\\\\) \\\\(a\\\\) test\\\\?"
}
