% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_df_lines.R
\name{gather_df_lines}
\alias{gather_df_lines}
\title{Concatenate text by a common key}
\usage{
gather_df_lines(
  df,
  key_colname,
  text,
  sep = " ",
  trimws = T,
  text_colname = "content"
)
}
\arguments{
\item{df}{A data.frame containing the data}

\item{key_colname}{\code{character} The column used as the grouping key (character string)}

\item{text}{\code{character} The column containing the text to be concatenated (character string)}

\item{sep}{\code{character} The separator used to concatenate the text (default is " ")}

\item{trimws}{\code{logical}, default = \code{TRUE} If \code{TRUE}, suppress leading, tail and double white spaces}

\item{text_colname}{\code{character} The colname for the concatened text column (default is "content")}
}
\value{
A data.frame with the text concatenated by key
}
\description{
This function concatenates text spread across multiple rows
based on a common key, using only base R functions.
}
\examples{
# Create a sample data.frame
#df <- data.frame(id = c(1, 1, 1, 2, 2, 3),
# text = c("Hello ", "world ", " !", "How", "are    you?", " Test "))
# Use the function to concatenate text by 'id'
# gather_df_lines(df, key = "id", text = "text")
}
