% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_filespath.R
\name{get_github_filespath}
\alias{get_github_filespath}
\title{Get GitHub File URLs Recursively Matching a Pattern}
\usage{
get_github_filespath(
  repo = "tidyverse/ggplot2",
  branch = "main",
  pattern = "\\\\.R",
  no_recursivity = F
)
}
\arguments{
\item{repo}{\strong{(character)} Either \code{"owner/repo"} or a full GitHub API URL.}

\item{branch}{\strong{(character)} The branch to scan for files. Default is \code{"main"}.}

\item{pattern}{\strong{(character)} A regex pattern to filter files (default is \code{"\\\\.R$"}, meaning R scripts).}

\item{no_recursivity}{\strong{logical} A boolean, used as internal value (prevent the func' to loop).}
}
\value{
A character vector of URLs pointing to the raw content of the matching files.
}
\description{
This function retrieves the full URLs of files in a GitHub repository that match a specified pattern,
searching recursively through all directories.
}
\examples{
\dontrun{
url1 <- get_github_filespath(repo = "tidyverse/ggplot2", branch = "main", pattern = "\\\\.R$")
urls <- get_github_filespath(repo = "https://api.github.com/repos/tidyverse/ggplot2/git/trees/main")
}
}
