% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_raw_filespath.R
\name{get_github_raw_filespath}
\alias{get_github_raw_filespath}
\title{Get GitHub File URLs Matching a Pattern}
\usage{
get_github_raw_filespath(
  repo = "tidyverse/ggplot",
  branch = "main",
  pattern = "\\\\.R"
)
}
\arguments{
\item{repo}{\strong{(character)} The GitHub repository in the format \code{"owner/repo"}. Default is \code{"tidyverse/tidyverse"}.}

\item{branch}{\strong{(character)} The branch to scan for files. Default is \code{"main"}.}

\item{pattern}{\strong{(character)} A regex pattern to filter files (default is \code{"\\\\.R"}, meaning R scripts).}
}
\value{
A character vector of URLs of the matching files (https://raw.githubusercontent.com/).
}
\description{
This function retrieves the full URLs of files in a GitHub repository that match a specified pattern.
}
\examples{
\dontrun{
readr_fp <- get_github_raw_filespath(repo = "tidyverse/readr", pattern = "\\\\.R")
}
}
