% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{print.corpus.lines}
\alias{print.corpus.lines}
\title{Print Method for corpus.lines}
\usage{
\method{print}{corpus.lines}(x, ...)
}
\arguments{
\item{x}{An object of class 'corpus.lines'.}

\item{...}{Additional parameters for the print function.}
}
\description{
Custom print method for objects of class 'corpus.lines'. Adds a header before printing the object.
}
