% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{print.corpus.nodelist}
\alias{print.corpus.nodelist}
\title{Print Method for corpus_lines}
\usage{
\method{print}{corpus.nodelist}(x, ...)
}
\arguments{
\item{x}{An object of class 'corpus.nodelist'.}

\item{...}{Additional parameters for the print function.}
}
\description{
Custom print method for objects of class 'corpus.nodelist'. Adds a header before printing the object.
}
\details{
Print a corpus.nodelist dataframe
}
