% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_commented_lines.R
\name{separate_commented_lines}
\alias{separate_commented_lines}
\title{Separate commented and non-commented lines from text}
\usage{
separate_commented_lines(texts, delim_pair = c(`/*` = "*/"), .verbose = T)
}
\arguments{
\item{texts}{\code{character} vector containing text with possible comments.}

\item{delim_pair}{\code{character} A named character vector of length 1, where the name is the opening delimiter
(e.g., \code{"/*"}) and the value is the closing delimiter (e.g., \code{"*/"}).}

\item{.verbose}{\code{logical}, default = \code{TRUE} Return a progress bar if set to \code{TRUE}}
}
\value{
A data frame with three columns:
\describe{
\item{text}{The original text.}
\item{comments}{The extracted comments, concatenated together if multiple are found.}
\item{codelines}{The remaining text after removing comments.}
}
}
\description{
This function extracts and separates commented and non-commented parts from a given vector of texts.
It uses a specified delimiter pair to identify comments and returns a data frame with the full text,
extracted comments, and remaining code lines.
}
\examples{
\dontrun{
texts <- c(
  "Code snippet /* first comment */ and more code /* second comment */",
  "Another line /* only a comment */",
  "A line with no comments", " here one start /* commented ", " some commented line", " */"
)

separate_commented_lines(texts, delim_pair = c("/*" = "*/"), .verbose = FALSE)
}
}
