% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srch_pattern_in_df.R
\name{srch_pattern_in_df}
\alias{srch_pattern_in_df}
\title{Try to match a text pattern in a df column by only extract the text}
\usage{
srch_pattern_in_df(
  df,
  content_col_name = "content",
  pattern = "(^| \\\\.|\\\\b)([\\\\.A-Za-z0-9_]+)(?=\\\\s*(?:<-)\\\\s*function)",
  match_to_exclude = NULL,
  ignore_match_less_than_nchar = 3,
  extracted_txt_col_name = "matches",
  duplicated_lines_are_normal = F
)
}
\arguments{
\item{df}{\code{data.frame}
A data.frame with a minima a \code{character} column.}

\item{content_col_name}{\code{character}, default = \code{"content"}
Name of the text column in the input df (will be returned in the output df).}

\item{pattern}{\code{character}, default = \code{"\\\\b([A-Za-z0-9_]+)(?=\\\\s*(?:<-|=)\\\\s*(?:function|$))"}
A regex for matching lines and extract text.}

\item{match_to_exclude}{\code{character} A vector of values that will not be returned such as a match.
The rows where the \code{values} match any element in this vector will be removed.}

\item{ignore_match_less_than_nchar}{\code{double}, default = 2 Excluding match depending on char. number of the matched text (strictly inferior)
Default exclude match of 1 char such as 'x'.}

\item{extracted_txt_col_name}{\code{character}, default = \code{"matches"}
Column name for the extracted text (last col' of the returned df)}

\item{duplicated_lines_are_normal}{\code{logical}, default = \code{FALSE}. If set to \code{TRUE}, silent the warning  about duplicated lines}
}
\value{
A \code{data.frame} similar to the one passed by the user with 1 more column : the match ; \emph{a minima} :
\describe{
\item{\code{content}}{\code{character} The text column designed by the user.}
\item{\code{match}}{\code{character} The matched text on this line, \code{NA} if there is no match.}
}
}
\description{
Read some files and answer the content readed in a df.
Then try to extract a pattern
and return the extracted text in a column of the returned df (NA meaning 'no match').
}
\seealso{
\code{\link{readlines_in_df}}
}
