% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_functions_list_to_corpus.R
\name{add_functions_list_to_corpus}
\alias{add_functions_list_to_corpus}
\title{Compute a function list from the codes df of a corpus.list object}
\usage{
add_functions_list_to_corpus(corpus, lang_dictionnary)
}
\arguments{
\item{corpus}{A \code{corpus.list} object with at least the 'codes' dataframe}

\item{lang_dictionnary}{\code{list} A language patterns dictionnary specific to one language}
}
\value{
a \code{list} of class \code{corpus.list} with exposed functions names and file path
}
\description{
Compute a function list from the codes df of a corpus.list object
}
