% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_corpus.R
\name{construct_corpus}
\alias{construct_corpus}
\title{Construct a list of DataFrames of Lines Readed From Files
from Local GitHub Repositories and/or Local Folders}
\usage{
construct_corpus(
  local_folders_paths = NULL,
  languages = "R",
  repos = NULL,
  ...
)
}
\arguments{
\item{local_folders_paths}{\code{character}. Default = \code{NULL}. A character vector of local folder paths to scan for code files.}

\item{languages}{\code{character}. Default = \code{"R"}. A character vector specifying the programming language(s) to include in the corpus.}

\item{repos}{\code{character}. Default = \code{NULL}. A character vector of GitHub repository URLs or repository identifiers to extract files from (e.g., \code{"user/repo"}).}

\item{...}{Additional arguments passed to \code{srch_pattern_in_files_get_df}
(filtering options, depth of folder scanning, names of the returned df columns, .verbose parameter, etc.).}
}
\value{
A list of 2 dataframes containing the corpus of collected files : \code{codes} and \code{comments}. The data frames typically includes columns such as:
\describe{
\item{\code{file_path}}{ \code{character} The local file path or constructed GitHub URL.}
\item{\code{line_number}}{\code{integer} The line number of the file.}
\item{\code{content}}{\code{character} The content in the line.}
}
}
\description{
Given a Language and a folder path(s) and/or github repo(s)
Build a list of 2 samely-structured dataframes containing files
from either GitHub repositories or local folder paths.
}
\details{
\itemize{
\item If \code{local_folders_paths} is provided (one or a list), the function scans the directories and retrieves file paths matching the specified languages.
\item If \code{repos} is provided (one or a list), it constructs URLs to the raw content of files from the specified GitHub repositories.
\item Both local paths and GitHub URLs can be combined in the final output.
}
}
\seealso{
\code{\link{readlines_in_df}}, \code{\link{get_github_raw_filespath}}, \code{\link{get_def_regex_by_language}}
}
