% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_constructor.R
\name{.construct.corpus.list}
\alias{.construct.corpus.list}
\title{Create a corpus.list from a list or append a dataframe to a corpus.list}
\usage{
.construct.corpus.list(corpus, df_to_add = NULL, ...)
}
\arguments{
\item{corpus}{A list of dataframes.}

\item{df_to_add}{A dataframe to append to the corpus.list.}

\item{...}{Additionnal attributes (e.g., keep trace of the languages analyzed).}
}
\value{
A dataframe of class 'corpus.list'.
}
\description{
Create a corpus.list from a list or append a dataframe to a corpus.list
}
