% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_networkd3_from_igraph.R
\name{get_networkd3_from_igraph}
\alias{get_networkd3_from_igraph}
\title{Convert an igraph object to a networkD3 visualization}
\usage{
get_networkd3_from_igraph(
  graph_igraph,
  title_h1 = "networkD3",
  subtitle_h2 = NULL,
  endnotes_h3 = NULL,
  colors_for_nodes = c("green", "grey", "black", "red"),
  color_outdeg_instead_of_indeg = F,
  color_for_na_link = colors_for_nodes[[1]],
  charge = -200,
  ...
)
}
\arguments{
\item{graph_igraph}{An \code{igraph} object representing the network.}

\item{title_h1}{\code{character}. The main title (HTML \verb{<h1>}) displayed above the graph. Default is \code{"networkD3"}.}

\item{subtitle_h2}{\code{character} (optional). The subtitle (HTML \verb{<h2>}) displayed below the main title. Default is \code{NULL} (no subtitle).}

\item{endnotes_h3}{\code{character} (optional). A description or note (HTML \verb{<h3>}) displayed after the network. Default is \code{NULL} (no text).}

\item{colors_for_nodes}{\code{character} vector. Specifies colors for nodes based on their degree and also number of separation
the func' will try to realize (based on indegrees or outdegrees). Default is \code{c("green", "grey", "black", "red")}.}

\item{color_outdeg_instead_of_indeg}{\code{logical}. If \code{TRUE}, node colors are assigned based on out-degree instead of in-degree. Default is \code{FALSE}.}

\item{color_for_na_link}{\code{character}. Color assigned to links when data is missing.
Default is the 1st color of \code{colors_for_node} (\code{"green"}).}

\item{charge}{\code{integer}, default = \code{-200} Numeric value indicating
either the strength of the node repulsion (negative value) or attraction (positive value).
Passed to \code{networkD3::forceNetwork}.}

\item{...}{optional parameter passed to \code{networkD3::forceNetwork}}
}
\value{
A "\code{forceNetwork}" & "\code{htmlwidget}" object, i.e. a list that symbolize an html object from \code{networkD3::forceNetwork}
}
\description{
This function transforms an \code{igraph} graph object into a \code{networkD3} interactive visualization.
It allows customization of titles, node colors, missing link colors, and additional HTML code insertion.
}
\examples{
invisible(library(igraph))
g <- make_ring(10)
get_networkd3_from_igraph(g, title_h1 = "My Network")
}
