% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readlines_in_df.R
\name{readlines_in_df}
\alias{readlines_in_df}
\title{Read text file with readLines and answer a df}
\usage{
readlines_in_df(
  files_path,
  return_lowered_text = FALSE,
  .verbose = TRUE,
  file_path_col_name = "file_path",
  content_col_name = "content",
  line_number_col_name = "line_number"
)
}
\arguments{
\item{files_path}{\code{character} Path(s) or url(s) to the text files to read.}

\item{return_lowered_text}{\code{logical}, default = \code{FALSE}
TRUE for lowercasing the returned text. FALSE (the default) will preserve the readed text}

\item{.verbose}{\code{logical}, default = \code{TRUE}
If \code{TRUE}, shows a progress bar while reading the file.}

\item{file_path_col_name}{\code{character}, default = \code{"file_path"}
Column name for the file path in the output dataframe.}

\item{content_col_name}{\code{character}, default = \code{"content"}
Column name for the file content in the output dataframe.}

\item{line_number_col_name}{\code{character}, default = \code{"line_number"}
Column name for the line numbers in the output dataframe.}
}
\value{
Return a \code{data.frame} with 3 columns.
The \emph{last} column (\code{content} by default) contain the readed lines from the file.
\describe{
\item{file_path}{\code{character} Path of the file where the match was found.}
\item{line_number}{\code{integer} Line number in the file.}
\item{content}{\code{character} Content of the line.}
}
}
\description{
This function reads the content of a text file line by line and stores it in a data frame.
You must pick up the FIRST col' for the file_path readed, and the LAST col for the content readed
intermediar are additionnal info, i.e. line_number
}
\examples{
\dontrun{
  df <- readlines_in_df("path/to/file.R", case.sensitive = TRUE)
  contents_readed <- df[[ncol(df)]]
  file_path_readed <-  df[[1]]
}
}
