% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_all_to_tidy_df.R
\name{str_extract_all_to_tidy_df}
\alias{str_extract_all_to_tidy_df}
\title{Extract regex matches from a string and return a tidy dataframe}
\usage{
str_extract_all_to_tidy_df(
  string,
  pattern,
  matches_colname = "matches",
  row_number_colname = "row_number"
)
}
\arguments{
\item{string}{\code{character} vector. A character vector containing the input text.}

\item{pattern}{\code{character}. A regex pattern to extract matches.}

\item{matches_colname}{\code{character}. A string specifying the column name for extracted matches (default: \code{"matches"}).}

\item{row_number_colname}{\code{character}. A string specifying the column name for row numbers (default: \code{"row_number"}).}
}
\value{
A dataframe with the extracted matches and their corresponding row numbers.
\describe{
\item{matches}{1st col' is the matched-text. Colname is indicated with the \code{matches_colname} parameter (default is 'matches')}
\item{row_number}{2nd col is the position of the match within the vector. Colname is indicated with the \code{row_number_colname} parameter (default is 'row_number')}
}
}
\description{
This function applies \code{stringr::str_extract_all()} to a string, extracts regex matches,
and returns a 2 columns unested dataframe :
1st column is the matched text. Lines without match are filtered out.
2nd column is the corresponding position index
Option offered : customizable colnames
}
\examples{
text_data <- c("Here is funcA and funcB", "Nothing here", "funcC is present")
pattern <- "func[A-C]"
result_df <- str_extract_all_to_tidy_df(text_data, pattern)
print(result_df)
}
